latest-news-crawler/requirements.txt
schedule
google-cloud-datastore
feedparser
news-please
tldextract